I-168: Apply ReceivableMessages format to several actors
