+++ examples/src/main/scala/examples/hybrid/api/http/WalletApiRoute.scala
@@ -9,7 +9,6 @@ import examples.hybrid.state.HBoxStoredState
 import examples.hybrid.wallet.HWallet
 import io.circe.parser._
 import io.circe.syntax._
-import scorex.core.LocalInterface.ReceivableMessages.LocallyGeneratedTransaction
 import scorex.core.api.http.{ApiException, ApiRouteWithFullView, SuccessApiResponse}
 import scorex.core.settings.RESTApiSettings
 import scorex.core.transaction.box.proposition.PublicKey25519Proposition
@@ -23,6 +22,8 @@ import scala.util.{Failure, Success, Try}
 case class WalletApiRoute(override val settings: RESTApiSettings, nodeViewHolderRef: ActorRef)
                          (implicit val context: ActorRefFactory)
   extends ApiRouteWithFullView[HybridHistory, HBoxStoredState, HWallet, SimpleBoxTransactionMemPool] {
+
+  import scorex.core.NodeViewHolder.ReceivableMessages.LocallyGeneratedTransaction
 
   //TODO move to settings?
   val DefaultFee = 100
