+++ examples/src/main/scala/examples/hybrid/mining/PowMiner.scala
@@ -7,9 +7,8 @@ import examples.hybrid.history.HybridHistory
 import examples.hybrid.state.HBoxStoredState
 import examples.hybrid.util.Cancellable
 import examples.hybrid.wallet.HWallet
-import scorex.core.LocalInterface.ReceivableMessages.LocallyGeneratedModifier
 import scorex.core.ModifierId
-import scorex.core.NodeViewHolder.{CurrentView, GetDataFromCurrentView}
+import scorex.core.NodeViewHolder.CurrentView
 import scorex.core.block.Block.BlockId
 import scorex.core.transaction.box.proposition.PublicKey25519Proposition
 import scorex.core.utils.ScorexLogging
@@ -29,6 +28,9 @@ import scala.util.Random
 class PowMiner(viewHolderRef: ActorRef, settings: HybridMiningSettings) extends Actor with ScorexLogging {
 
   import PowMiner._
+  import scorex.core.NodeViewHolder.ReceivableMessages.{LocallyGeneratedModifier, GetDataFromCurrentView}
+
+
 
   private var cancellableOpt: Option[Cancellable] = None
   private var mining = false
