+++ examples/src/main/scala/examples/hybrid/wallet/SimpleBoxTransactionGenerator.scala
@@ -5,8 +5,7 @@ import examples.commons.{SimpleBoxTransaction, SimpleBoxTransactionMemPool}
 import examples.commons.Value
 import examples.hybrid.history.HybridHistory
 import examples.hybrid.state.HBoxStoredState
-import scorex.core.LocalInterface.ReceivableMessages.LocallyGeneratedTransaction
-import scorex.core.NodeViewHolder.{CurrentView, GetDataFromCurrentView}
+import scorex.core.NodeViewHolder.CurrentView
 import scorex.core.transaction.box.proposition.PublicKey25519Proposition
 import scorex.core.utils.ScorexLogging
 
@@ -20,7 +19,8 @@ import scala.util.{Failure, Random, Success, Try}
   */
 class SimpleBoxTransactionGenerator(viewHolderRef: ActorRef) extends Actor with ScorexLogging {
 
-  import SimpleBoxTransactionGenerator._
+  import SimpleBoxTransactionGenerator.ReceivableMessages._
+  import scorex.core.NodeViewHolder.ReceivableMessages.{GetDataFromCurrentView, LocallyGeneratedTransaction}
 
   private val getRequiredData: GetDataFromCurrentView[HybridHistory,
     HBoxStoredState,
@@ -70,11 +70,10 @@ class SimpleBoxTransactionGenerator(viewHolderRef: ActorRef) extends Actor with 
 }
 
 object SimpleBoxTransactionGenerator {
-
-  case class StartGeneration(delay: FiniteDuration)
-
-  case class GeneratorInfo(tx: Try[SimpleBoxTransaction])
-
+  object ReceivableMessages {
+    case class StartGeneration(delay: FiniteDuration)
+    case class GeneratorInfo(tx: Try[SimpleBoxTransaction])
+  }
 }
 
 object SimpleBoxTransactionGeneratorRef {
