+++ testkit/src/main/scala/scorex/testkit/properties/NodeViewHolderTests.scala
@@ -3,7 +3,6 @@ package scorex.testkit.properties
 import akka.actor._
 import org.scalatest.{Matchers, PropSpec}
 import org.scalatest.prop.PropertyChecks
-import scorex.core.LocalInterface.ReceivableMessages.LocallyGeneratedModifier
 import scorex.core.NodeViewHolder.EventType._
 import scorex.core.consensus.{History, SyncInfo}
 import scorex.core.transaction.box.proposition.Proposition
@@ -54,13 +53,14 @@ MPool <: MemoryPool[TX, MPool]]
   val syntactic = Seq(SuccessfulSyntacticallyValidModifier, SyntacticallyFailedPersistentModifier)
   val allEvents = semantic ++ syntactic
 
-  import NodeViewHolder._
+  import NodeViewHolder.ReceivableMessages._
+  import NodeViewHolder.ReceivableMessages.{Subscribe, GetDataFromCurrentView, LocallyGeneratedModifier}
 
 
   property("NodeViewHolder syntactically valid modifier subscription") { withFixture { ctx =>
     import ctx._
 
-    node ! NodeViewHolder.Subscribe(syntactic)
+    node ! Subscribe(syntactic)
     node ! GetDataFromCurrentView[HT, ST, Vault[P, TX, PM, _], MPool, PM] { v => totallyValidModifiers(v.history, v.state, 2).head }
     val mod = receiveOne(5 seconds).asInstanceOf[PM]
     node ! LocallyGeneratedModifier(mod)
@@ -70,7 +70,7 @@ MPool <: MemoryPool[TX, MPool]]
   property("NodeViewHolder: syntactically failed modifier subscription") { withFixture { ctx =>
     import ctx._
 
-    node ! NodeViewHolder.Subscribe(syntactic)
+    node ! Subscribe(syntactic)
     val invalid = syntacticallyInvalidModifier(h)
     node ! LocallyGeneratedModifier(invalid)
     expectMsgType[SyntacticallyFailedModification[PM]]
@@ -79,7 +79,7 @@ MPool <: MemoryPool[TX, MPool]]
   property("NodeViewHolder: semantically valid modifier subscription") { withFixture { ctx =>
     import ctx._
 
-    node ! NodeViewHolder.Subscribe(allEvents)
+    node ! Subscribe(allEvents)
     node ! GetDataFromCurrentView[HT, ST, Vault[P, TX, PM, _], MPool, PM] { v => totallyValidModifiers(v.history, v.state, 2).head }
     val mod = receiveOne(5 seconds).asInstanceOf[PM]
     node ! LocallyGeneratedModifier(mod)
@@ -90,7 +90,7 @@ MPool <: MemoryPool[TX, MPool]]
   property("NodeViewHolder: semantically failed modifier subscription") { withFixture { ctx =>
     import ctx._
 
-    node ! NodeViewHolder.Subscribe(allEvents)
+    node ! Subscribe(allEvents)
     node ! GetDataFromCurrentView[HT, ST, Vault[P, TX, PM, _], MPool, PM] { v => semanticallyInvalidModifier(v.state) }
     val invalid = receiveOne(5 seconds).asInstanceOf[PM]
     node ! LocallyGeneratedModifier(invalid)
@@ -101,7 +101,7 @@ MPool <: MemoryPool[TX, MPool]]
   property("NodeViewHolder: syntactically/semantically valid modifier subscription") { withFixture { ctx =>
     import ctx._
 
-    node ! NodeViewHolder.Subscribe(allEvents)
+    node ! Subscribe(allEvents)
 
     node ! GetDataFromCurrentView[HT, ST, Vault[P, TX, PM, _], MPool, PM] { v => totallyValidModifiers(v.history, v.state, 2).head }
     val mod = receiveOne(5 seconds).asInstanceOf[PM]
@@ -128,7 +128,7 @@ MPool <: MemoryPool[TX, MPool]]
 
   property("NodeViewHolder: check that valid modifiers are applicable") { withFixture { ctx =>
     import ctx._
-    node ! NodeViewHolder.Subscribe(Seq(SuccessfulSyntacticallyValidModifier, SyntacticallyFailedPersistentModifier))
+    node ! Subscribe(Seq(SuccessfulSyntacticallyValidModifier, SyntacticallyFailedPersistentModifier))
 
     node ! GetDataFromCurrentView[HT, ST, Vault[P, TX, PM, _], MPool, Seq[PM]] { v =>
       totallyValidModifiers(v.history, v.state, 10) //todo: fix magic number
@@ -144,7 +144,7 @@ MPool <: MemoryPool[TX, MPool]]
 
   property("NodeViewHolder: apply locally generated mod") { withFixture { ctx =>
     import ctx._
-    node ! NodeViewHolder.Subscribe(Seq(SuccessfulSyntacticallyValidModifier, SyntacticallyFailedPersistentModifier))
+    node ! Subscribe(Seq(SuccessfulSyntacticallyValidModifier, SyntacticallyFailedPersistentModifier))
 
     val invalid = syntacticallyInvalidModifier(h)
 
@@ -168,7 +168,7 @@ MPool <: MemoryPool[TX, MPool]]
 
     val waitDuration = 5.seconds
 
-    node ! NodeViewHolder.Subscribe(Seq(SuccessfulSyntacticallyValidModifier, SyntacticallyFailedPersistentModifier))
+    node ! Subscribe(Seq(SuccessfulSyntacticallyValidModifier, SyntacticallyFailedPersistentModifier))
 
     node ! GetDataFromCurrentView[HT, ST, Vault[P, TX, PM, _], MPool, Seq[PM]] { v => totallyValidModifiers(v.history, v.state, 2) }
     val initMods = receiveOne(waitDuration).asInstanceOf[Seq[PM]]
